# openapi.yaml
openapi: '3.0.2'
info:
  title: SCS Status Page API
  version: '1.0'
components:
  schemas:
    Id:
      type: string
    IncidentImpactType:
      type: string
    IncidentPhase:
      type: string
    Labels:
      type: object
      additionalProperties:
        type: string
    Component:
      type: object
      required:
        - id
        - displayName
        - labels
        - affectedBy
      properties:
        id:
          type: string
        displayName:
          type: string
        labels:
          $ref: '#/components/schemas/Labels'
        affectedBy:
          type: array
          items:
            $ref: '#/components/schemas/Id'
    Incident:
      type: object
      required:
        - id
        - title
        - affects
        - impactType
        - phase
      properties:
        id:
          type: string
        title:
          type: string
        affects:
          type: array
          items:
            $ref: '#/components/schemas/Id'
        beganAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        impactType:
          $ref: '#/components/schemas/IncidentImpactType'
        phase:
          $ref: '#/components/schemas/IncidentPhase'
paths:
  /phases:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncidentPhase'
  /impacttypes:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncidentImpactType'
  /components/{componentId}:
    get:
      summary: get specific component by id
      operationId: getComponent
      parameters:
      - in: path
        name: componentId
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Component'
  /components:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Component'
  /incident/{incidentId}:
    get:
      summary: Get specific incident by id
      operationId: getIncident
      parameters:
      - in: path
        name: incidentId
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
  /incidents:
    get:
      summary: Get list of incidents
      parameters:
      - in: query
        name: start
        schema:
          type: string
          format: date-time
        required: true
        description: Start of time frame to query for (RFC3339)
      - in: query
        name: end
        schema:
          type: string
          format: date-time
        required: true
        description: End of time frame to query for (RFC3339)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Incident'